+++ src/Attribute/RouteAttributeInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Router\Attribute;
+
+use Yiisoft\Router\Route;
+
+interface RouteAttributeInterface
+{
+    public function getRoute(): Route;
+}