+++ src/RouteCollector.php
@@ -4,12 +4,19 @@ declare(strict_types=1);
 
 namespace Yiisoft\Router;
 
+use Yiisoft\Router\Provider\RoutesProviderInterface;
+
 final class RouteCollector implements RouteCollectorInterface
 {
     /**
      * @var Group[]|Route[]
      */
     private array $items = [];
+
+    /**
+     * @var RoutesProviderInterface[]
+     */
+    private array $providers = [];
 
     /**
      * @var array[]|callable[]|string[]
@@ -21,6 +28,15 @@ final class RouteCollector implements RouteCollectorInterface
         array_push(
             $this->items,
             ...array_values($routes),
+        );
+        return $this;
+    }
+
+    public function addProvider(RoutesProviderInterface ...$provider): RouteCollectorInterface
+    {
+        array_push(
+            $this->providers,
+            ...array_values($provider),
         );
         return $this;
     }
@@ -45,6 +61,12 @@ final class RouteCollector implements RouteCollectorInterface
 
     public function getItems(): array
     {
+        foreach ($this->providers as $provider) {
+            array_push(
+                $this->items,
+                ...$provider->getRoutes(),
+            );
+        }
         return $this->items;
     }